feat: add Playground struct
This struct is a generalized way of extracting and storing information from context about the environment where the command has been made.
